Move converters to file; NIX default params
All the pyany_to_* were moved to a new file to reduce clutter in
the bocpd file. Also the NormalInvChiSquare prior didn't have
default params like the NormalGamma and NormalInvGamma, so I added
them for consistency.
